+++ internal/proxies/anthropic/usage_test.go
@@ -21,7 +21,7 @@ func TestTracker_AddAccumulates(t *testing.T) {
 	assert.Equal(t, int64(150), model.InputTokens)
 	assert.Equal(t, int64(225), model.OutputTokens)
 	// 150 input @ $3/MTok + 225 output @ $15/MTok = 0.00045 + 0.003375 = 0.003825
-	assert.InDelta(t, 0.003825, snap.CostUSD, 1e-9)
+	assert.InDelta(t, 0.003825, float64(snap.CostUSD), 1e-9)
 }
 
 func TestTracker_WritesUsageJSONAtomically(t *testing.T) {
@@ -113,7 +113,7 @@ func TestSSEInterceptor_IgnoresGarbage(t *testing.T) {
 	_, err := io.Copy(io.Discard, w)
 	require.NoError(t, err)
 	require.NoError(t, w.Close())
-	assert.InDelta(t, 0.0, tr.snapshot().CostUSD, 1e-9)
+	assert.InDelta(t, 0.0, float64(tr.snapshot().CostUSD), 1e-9)
 }
 
 type nopReadCloser struct{ io.Reader }
